# backend/docs/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Localis Data API",
        "description": "This is a simple Express API based on the supplied Localis data. It currently supports retrieving all data combine and also data in the 2 original formats - length of stay data and occupancy data",
        "version": "1.0.0",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "j.hogan@qut.edu.au"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
    "paths": {
...
        "/users/login": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Logs user into the system",
                "description": "",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "description": "User's email address",
                                        "example": "j.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "description": "User's password",
                                        "example": "123password"
                                    }
                                },
                                "required": [
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "schema": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            }
                        },
                        "content": {
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "User registers to become a user of the system",
                "description": "",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "description": "User's email address",
                                        "example": "j.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "description": "User's password",
                                        "example": "123password"
                                    },
                                    "LGAName": {
                                        "type": "string",
                                        "description": "Local Government Area of the user",
                                        "example": "Noosa"
                                    }
                                },
                                "required": [
                                    "email",
				    "password",
                                    "LGAName"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "schema": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            }
                        },
                        "content": {
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        }
    }
}